backend/requirements.txt
mysql-connector-python
appdaemon
